test(ci): centralize aggregator boot auth helper and document fixes
